shadow-dom/events/event-dispatch/test-003.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_05_05_03</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="https://w3c.github.io/webcomponents/spec/shadow/#event-path-trimming">
<meta name="assert" content="Event Path Trimming: event listeners for trusted events must not be invoked on a node for which the target and relatedTarget are the same.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
var A_05_05_03_T01 = async_test('A_05_05_03_T01');

A_05_05_03_T01.step(unit(function (ctx) {

    var d = newRenderedHTMLDocument(ctx);

    var host = d.createElement('div');
    host.setAttribute('id', 'host');
    d.body.appendChild(host);

    //Shadow root to play with
    var s = host.createShadowRoot();

    var input1 = d.createElement('input');
    input1.setAttribute('id', 'input1');
    s.appendChild(input1);

    var input2 = d.createElement('input');
    input2.setAttribute('id', 'input2');
    s.appendChild(input2);

    // Focus before installing the event handler.
    input2.focus();

    // Add an event listner above its shadow root. If both target and
    // relatedTarget are in the same shadow tree under the root,
    // retargeted target and relatedTarget will be the same node
    // and the event must not invoke the handler.
    host.addEventListener('focusin', A_05_05_03_T01.step_func(function(event) {
        if (('isTrusted' in event) && event.isTrusted &&
            (event.target == event.relatedTarget)) {
            assert_true(false, 'Trusted event must not invoke its listeners' +
                'if target and relatedTarget are the same');
        }
        if (event.target !== event.relatedTarget)
            return;
        assert_true(false, 'Cannot determine whether the event is trusted, ' +
            'but event handler is invoked anyway with ' +
            'target === relatedTarget.');
    }), false);

    // Causes a "focusin" event, from #input2 to #input1.
    // It should be viewed outside the shadow as "target == relatedTarget".
    input1.focus();

    A_05_05_03_T01.done();
}));
</script>
</body>
</html>
